chore(karma): move karma config to config/karma and update to use webpack
