+++ config/karma/karma-ci.config.js
@@ -1,0 +1,1 @@
+module.exports = require('./karma.config').ci
